refactor: extract remaining gaussian streaming type headers
